new base class for integration tests that use REST
