+++ src/test/java/uk/ac/cam/cl/dtg/isaac/api/EventsFacadeIT.java
@@ -17,7 +17,6 @@ import java.sql.PreparedStatement;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Collectors;
